+++ packages/jimp/package.json
@@ -70,6 +70,7 @@
     "@jimp/plugin-mask": "0.3.5",
     "@jimp/plugin-gaussian": "0.3.5",
     "@jimp/plugin-scale": "0.3.5",
+    "@jimp/plugin-displace": "0.3.5",
     "@jimp/plugin-invert": "0.3.5",
     "@jimp/plugin-crop": "0.3.5"
   },
