+++ packages/plugin-displace/README.md
@@ -1,0 +1,3 @@
+# @jimp/plugin-displace
+
+displace an image.
